# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pulse Monitor API
  description: >
    Pulse Monitor API is a RESTful uptime monitoring service that continuously
    checks websites, records time-series monitoring data, manages incidents,
    and exposes a public status page.

  version: 1.0.0
servers:
  - url: https://pulse-monitor-yd3n.onrender.com/
    description: Production
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false

        message:
          type: string
          example: Validation failed
...
    Monitor:
      type: object

      properties:
        id:
          type: integer
          example: 1

        owner_id:
          type: integer
          example: 1

        name:
          type: string
          example: Google

        url:
          type: string
          example: https://google.com

        interval_seconds:
          type: integer
          example: 60

        expected_status:
          type: integer
          example: 200

        is_active:
          type: boolean
          example: true

        created_at:
          type: string
          format: date-time

        ok:
          type: boolean
          nullable: true
          example: true

        status_code:
          type: integer
          nullable: true
          example: 200

        checked_at:
          type: string
          format: date-time
          nullable: true
...
    UpdateMonitor:
      type: object

      properties:
        name:
          type: string
          example: Google API

        url:
          type: string
          example: https://google.com

        interval_seconds:
          type: integer
          example: 30

        expected_status:
          type: integer
          example: 200

        is_active:
          type: boolean
          example: false
...
paths:
...
  /monitors/{id}:
    get:
      tags:
        - Monitors
      summary: Get a monitor
      security:
        - bearerAuth: []

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '200':
          description: Monitor found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true

                  data:
                    $ref: '#/components/schemas/Monitor'

        '401':
          description: Unauthorized

        '404':
          description: Monitor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Monitors
      summary: Update a monitor
      security:
        - bearerAuth: []

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMonitor'

      responses:
        '200':
          description: Monitor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'

        '400':
          description: Validation failed

        '401':
          description: Unauthorized

        '404':
          description: Monitor not found

        '500':
          description: Internal server error

    delete:
      tags:
        - Monitors
      summary: Delete a monitor
      security:
        - bearerAuth: []

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        '204':
          description: Monitor deleted

        '401':
          description: Unauthorized

        '404':
          description: Monitor not found
...